hdfs.py refactoring: Move out clients
