+++ test/minicluster.py
@@ -52,8 +52,11 @@ class MiniClusterTestCase(unittest.TestCase):
         self.fs = luigi.contrib.hdfs.client
         cfg_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "testconfig")
         hadoop_bin = os.path.join(os.environ['HADOOP_HOME'], 'bin/hadoop')
-        # TODO: Remove this hacky monkey patching (@tarrasch 29-Apr-2015)
-        luigi.contrib.hdfs.load_hadoop_cmd = lambda: [hadoop_bin, '--config', cfg_path]
+        # TODO: Remove this hacky patching (@tarrasch 29-Apr-2015)
+        cmd = lambda: [hadoop_bin, '--config', cfg_path]
+        luigi.contrib.hdfs.load_hadoop_cmd = cmd
+        luigi.contrib.hdfs.clients.load_hadoop_cmd = cmd
+        luigi.contrib.hdfs.format.load_hadoop_cmd = cmd
 
     def tearDown(self):
         if self.fs.exists(self._test_dir()):
